pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-rec-platform"
version = "0.1.0"
description = "融合 LLM 的智能推荐系统平台"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Web 框架
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "starlette>=0.36.0",

    # 数据校验
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # 配置
    "pyyaml>=6.0",

    # 存储
    "redis>=5.0.0",
    "pymysql>=1.1.0",
    "clickhouse-driver>=0.2.6",

    # 向量检索
    "faiss-cpu>=1.7.4",

    # 深度学习
    "torch>=2.2.0",
    "lightgbm>=4.2.0",

    # LLM Agent
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "openai>=1.12.0",

    # 监控
    "prometheus-client>=0.20.0",

    # 序列化
    "msgpack>=1.0.7",
    "protobuf>=4.25.0",
    "grpcio>=1.60.0",

    # 数据处理
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "grpcio-tools>=1.60.0",
]

[tool.setuptools.packages.find]
include = [
    "configs*",
    "protocols*",
    "server*",
    "pipeline*",
    "llm*",
    "feature*",
    "monitor*",
    "storage*",
    "utils*",
    "scripts*",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
